edit in database

+++ database/migrations/2023_08_16_113207_create_user_information_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('user_information', function (Blueprint $table) {
             $table->id();
-            $table->integer('civil_number');
+            $table->bigInteger('civil_number');
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
